+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 #0.4.0
-add method message
-add option newestOnTop
-add mobile style
+####add method message
+####add option newestOnTop
+####add mobile style
 
 #0.3.1
 Nothing new just some problems with npm readme.md file